requirements.txt
edge-tts==7.2.7
requests===2.32.5
python-dotenv==1.2.1
openai-whisper==20250625
google-api-python-client==2.188.0
google-auth-oauthlib===1.2.4
